Added a logo, title, tagline and laid it out in a grid. Moved the authentication to links that are pinned to the top right of the page and when a user logsout then the nav menu disappears and says Not Authenticated and in the 3rd cell it says Please sign in.
