+++ views/frontEnd/src/assets/css/mediaQueries.css
@@ -12,60 +12,79 @@
  */
 body {
    font-family: 'Poppins', sans-serif;
+   font-size: 1.9rem; 
+}
+.wrapper {
+   display: grid;
+   grid-template-columns: repeat(9, [col-start] 1fr);
+   grid-auto-rows: minmax(100px, auto);
+   grid-template-areas:
+     "hd hd hd hd   hd   hd   hd   hd   hd"
+     "sd sd sd main main main main main main"
+     "sd sd sd  ft  ft   ft   ft   ft   ft";
+}
+/** Header **/
+header {
+   display: grid;
+   grid-area: hd;
+   grid-column: 1 / span 9; /* Span from the first column to the last column */
+   justify-content: center; /* Align the header horizontally */
+   align-items: center; /* Align the header vertically */
+   gap: 0; /* Adjust the gap between logo and sub-header */
+   width: 100%; /* Ensure the header spans the entire width of the parent container */
+   padding-bottom: 10px;
+   box-shadow: 0 0 4px 0 rgba(99, 89, 71, 0.08), 0 2px 4px 0 rgba(68, 58, 43, 0.12);
 }
 
-button {
-   font-size: 85%;
-}
-.btn {
-   background-color: var(--background-color);
-   color: #505050;
-   padding: .5em 1em;
-   border: none;
-   outline: none;
-   position: relative;
-   cursor: pointer;
-
-   --background-color: #D9D9D9;
-   --border-color: #A5A5A5;
-   --border-size: 1px;
+.sub-header {
+   grid-column: col-start 3 / col-start 8; /* Start from col-start 3 and end at col-start 9 */
+   display: flex; /* Use flexbox */
+   flex-direction: column; /* Arrange items vertically */
+   justify-content: center; /* Center items horizontally */
+   align-items: center; /* Center items vertically */
+   gap: 5px; /* Adjust the gap between elements within sub-header */
 }
 
-.btn.btn-border-pop::before {
-   content: '';
-   position:absolute;
-   top:0;
-   bottom:0;
-   right:0;
-   left:0;
-   z-index: -2;
-   border: var(--border-size) solid var(--border-color);
-   transition: top, left, right, bottom, 100ms ease-in-out;
-   border-radius: 4px;
+header .logo {
+   grid-column: col-start -1 / span 2;
+   justify-items: center;
+   width: 300px;
 }
 
-
-.btn.btn-border-pop:hover::before,
-.btn.btn-border-pop:focus::before {
-   top: calc(var(--border-size) * -2);
-   left: calc(var(--border-size) * -2);
-   right: calc(var(--border-size) * -2);
-   bottom: calc(var(--border-size) * -2);
+header h1 {
+   color: #006358;
+   font-size: 18px;
 }
 
-.btn.large-button {
-   font-size: 14px;
-   border-radius: 4px;
+nav {
+   grid-column: col-start 1 / col-start 3;
 }
-.btn.square {
-   width: 100%;
-   height: 0;
-   padding-top: 100%;
+
+.auth {
+   position: fixed;
+   top: 10px; /* Adjust the top position as needed */
+   right: 10px; /* Adjust the right position as needed */
 }
-button {
-   display: flex;
+
+.auth a {
+   text-decoration: none;
+   font-size:medium;
 }
- 
+
+.content {
+   grid-area: main;
+}
+.sidebar {
+   grid-area: sd;
+}
+.footer {
+   grid-area: sd
+}
+
+/* .sub-wrapper img.bars {
+   width: 600px;
+   height: 20px;
+} */
 
 @media all and (min-width: 1024px) and (max-width: 1280px) {
    * {margin: 0; padding: 0;}
@@ -73,7 +92,17 @@ button {
 
 
 
- @media all and (min-width: 768px) and (max-width: 1024px) {* {margin: 0; padding: 0;}}
+ @media all and (min-width: 768px) and (max-width: 1024px) {
+   .header {
+      font-size: 1.0rem; 
+   }
+   .header img{
+      width: 200px;
+   }
+   
+
+   * {margin: 0; padding: 0;}
+}
 
  @media all and (min-width: 480px) and (max-width: 768px) {* {margin: 0; padding: 0;}}
 
@@ -81,9 +110,15 @@ button {
 
 
  /* Portrait */
- @media screen and (orientation:portrait) { /* Portrait styles here */ * {margin: 0; padding: 0;} }
+ @media screen and (orientation:portrait) { 
+   /* Portrait styles here */ 
+   * {margin: 0; padding: 0;}
+}
  /* Landscape */
- @media screen and (orientation:landscape) { /* Landscape styles here */ * {margin: 0; padding: 0;}}
+ @media screen and (orientation:landscape) { 
+   /* Landscape styles here */ 
+   * {margin: 0; padding: 0;}
+}
 
 
  /* CSS for iPhone, iPad, and Retina Displays */
@@ -104,7 +139,11 @@ button {
  @media screen and (max-device-width: 480px) and (orientation:landscape) {* {margin: 0; padding: 0;}}
 
  /* iPad Portrait */
- @media screen and (min-device-width: 481px) and (orientation:portrait) {* {margin: 0; padding: 0;}}
+@media screen and (min-device-width: 481px) and (orientation:portrait) {
+   * {margin: 0; padding: 0;}
+}
 
  /* iPad Landscape */
- @media screen and (min-device-width: 481px) and (orientation:landscape) {* {margin: 0; padding: 0;}}
+ @media screen and (min-device-width: 481px) and (orientation:landscape) {
+   * {margin: 0; padding: 0;}
+}
