+++ views/frontEnd/src/assets/css/reset.css
@@ -56,21 +56,14 @@ input[type="text"] {
 }
 body {
   width: 100%;
-  max-width: 68rem;
   margin: 0 auto;
-  font:
-    1.6rem/1.25 "Helvetica Neue",
-    Helvetica,
-    Arial,
-    sans-serif;
-  background-color: #f5f5f5;
+  background-color: #ffffff;
   color: #4d4d4d;
   -moz-osx-font-smoothing: grayscale;
   -webkit-font-smoothing: antialiased;
 }
 @media screen and (min-width: 620px) {
   body {
-    font-size: 1.9rem;
     line-height: 1.31579;
   }
 }
