+++ views/frontEnd/src/components/Header.vue
@@ -1,0 +1,15 @@
+<template>
+    <header>
+        <img class="logo" src="../assets/images/acto_pe_logo_400.png" alt="P&E Metrics and tools logo"/>
+        <article class="sub-header">
+            <h1>ACTO P&E Metrics and Tools</h1>
+            <img class="bars" src="../assets//images/bars_light.png" alt="colored bars that seperate text">
+            <p>A set of P&E metrics and helper tools</p>
+        </article>
+        <Auth class="auth" />
+    </header>
+</template>
+
+<script setup>
+   import Auth from '../components/Auth.vue'
+</script>
