+++ views/frontEnd/src/main.js
@@ -4,8 +4,8 @@ import { createApp } from 'vue';
 import App from './App.vue';
 import router from '../router/index.js';
 import { createPinia } from 'pinia';
-// import './assets/css/reset.css'
-// import './assets/css/mediaQueries.css'
+import './assets/css/reset.css'
+import './assets/css/mediaQueries.css'
 
 createApp(App)
   .use(createPinia())
